requirements.txt
openai
pypdf
python-dotenv
numpy
tiktoken


